requirements.txt
Flask==2.1.1
pymongo==3.12.1
flask_pymongo==2.3.0
Jinja2==3.1.0
PyPDF2
pdf2image
langchain
huggingface_hub
openai
chromadb
tiktoken
pickle5
unstructured[csv]
tabulate
sentence_transformers
bitsandbytes
accelerate
torch
pgvector
psycopg
sqlalchemy
sqlalchemy-utils
psycopg2
einops
python-dotenv
pytesseract
numpy
